Standarized cancellation from NumericalChoiceMenus
Every time a NumericalChoiceMenu is run() within a method, its return
value must be checked as when it's None or False the user has chosen
to cancel the menu and the method must return True to go back to the
previous menu (and not skip it on the way back).

Having to return True manually might be a bit hard on the coding
process but it seems to be a somewhat readable consistency. We're
currently assessing whether to return a sort of MenuResponse object
of sorts that would manage this behavior, but it is not a prioritary
concern.

We now add a 'Return' item to every NumericalChoiceMenu by default,
switching the opt-in method _add_return_option to the opt-out method
_remove_quit_item as this one will only be used in Program Entry menus
where the default quit behavior might be overriden as
necessary (instead of returning the typical lambda: False).