Initial implementation of battle system
We're able to execute a battle loop without crashing (no, seriously!)
so we're getting somewhere with this.

The Entity class wraps a planet or an archmage into an object with
*individualized* soldiers and mages, i.e, with health and attack
 values persistent through a playing session within the galaxy. Upon
 returning to the Main Menu, these persistences are lost as there is
 no way to recover some of these values given the file saving
 specifications.

We have *not* yet implemented: writing changes to the planet object (I
think), cleanup of turn information (getting rid of negative values
and the negative attack?), the implementation of a few abilities, and
code cleanup in general (lots of nested dots in here. "There must be a
better way!")